2025-07-08 level 2~3 과제 진행

+++ Calculator-lv1/src/Level1/Models/request_lib.java
@@ -3,6 +3,6 @@ package Level1.Models;
 public class request_lib {
     public String strException = "";
     public boolean resultBool = true;
-    public int resultInt;
+    public double resultInt;
 }
